# Set maxElapsedMillis on HttpEventPublisher to speedup Splunk tests